Compsognathus bodysize fix
The body size of Compsognathus is currently set to 1.5. Compare that to the human body size of 1.0 and realize that this would make them quite a bit larger than the average human.

But in real life Compsognathus was one of the smallest dinosaurs, standing at most little over 3 feet tall. They are also portrayed this way in the Jurassic Park movies. 

I've set the body size to 0.5 now; which should more accurately reflect their size in both reality and the movies. I've also updated their MaxPreyBodySize so they won't hunt things that would now chomp them in half with ease.